networking-switch-demo/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-shell-networking-standard==5.0.4
# cloudshell-shell-connectivity-flow>=1.1,<2.0
cloudshell-cli==4.0.1